requirements.txt
# Core ML and Data Processing
tensorflow>=2.13.0,<2.16.0
numpy>=1.21.0,<1.25.0
pandas>=1.5.0,<2.1.0
pillow>=9.0.0,<11.0.0
opencv-python>=4.7.0,<4.9.0

# Web Framework
flask>=2.2.0,<3.0.0
werkzeug>=2.2.0,<3.0.0

# Dataset Download (optional)
kaggle>=1.5.0,<1.6.0

# Image Processing and Visualization
matplotlib>=3.6.0,<3.8.0
scikit-learn>=1.2.0,<1.4.0
seaborn>=0.11.0,<0.13.0